ComradeBase

Documentation for ComradeBase.

ComradeBase.AbstractModel

ComradeBase.AbstractPolarizedModel

ComradeBase.CirBasis

ComradeBase.CoherencyMatrix

ComradeBase.CoherencyMatrix

ComradeBase.CoherencyMatrix

ComradeBase.CoherencyMatrix

ComradeBase.CoherencyMatrix

ComradeBase.DensityAnalytic

ComradeBase.ElectricFieldBasis

ComradeBase.GriddedKeys

ComradeBase.GriddedKeys

ComradeBase.GriddedKeys

ComradeBase.IntensityMap

ComradeBase.IsAnalytic

ComradeBase.IsPrimitive

ComradeBase.LPol

ComradeBase.LinBasis

ComradeBase.NotAnalytic

ComradeBase.NotPrimitive

ComradeBase.PolBasis

ComradeBase.PrimitiveTrait

ComradeBase.RPol

ComradeBase.SingleStokes

ComradeBase.StokesIntensityMap

ComradeBase.StokesParams

ComradeBase.XPol

ComradeBase.YPol

ComradeBase.axisdims

ComradeBase.basis_components

ComradeBase.basis_transform

ComradeBase.centroid

ComradeBase.dims

ComradeBase.evpa

ComradeBase.fieldofview

ComradeBase.flux

ComradeBase.header

ComradeBase.header

ComradeBase.imagegrid

ComradeBase.imagepixels

ComradeBase.imanalytic

ComradeBase.innerprod

ComradeBase.intensity_point

ComradeBase.intensitymap

ComradeBase.intensitymap

ComradeBase.intensitymap

ComradeBase.intensitymap!

ComradeBase.intensitymap!

ComradeBase.ispolarized

ComradeBase.isprimitive

ComradeBase.linearpol

ComradeBase.m̆

ComradeBase.named_dims

ComradeBase.phasecenter

ComradeBase.pixelsizes

ComradeBase.radialextent

ComradeBase.second_moment

ComradeBase.second_moment

ComradeBase.stokes

ComradeBase.visanalytic

ComradeBase.visibility_point

ComradeBase.AbstractModel — Type

AbstractModel

The Comrade abstract model type. To instantiate your own model type you should subtybe from this model. Additionally you need to implement the following methods to satify the interface:

Mandatory Methods

isprimitive: defines whether a model is standalone or is defined in terms of other models. is the model is primitive then this should return IsPrimitive() otherwise it returns NotPrimitive()

visanalytic: defines whether the model visibilities can be computed analytically. If yes then this should return IsAnalytic() and the user must to define visibility_point. If not analytic then visanalytic should return NotAnalytic().

imanalytic: defines whether the model intensities can be computed pointwise. If yes

then this should return IsAnalytic() and the user must to define intensity_point. If not analytic then imanalytic should return NotAnalytic().

radialextent: Provides a estimate of the radial extent of the model in the image domain. This is used for estimating the size of the image, and for plotting.

flux: Returns the total flux of the model.

Optional Methods:

intensity_point: Defines how to compute model intensities pointwise. Note this is must be defined if imanalytic(::Type{YourModel})==IsAnalytic().

visibility_point: Defines how to compute model visibilties pointwise. Note this is must be defined if visanalytic(::Type{YourModel})==IsAnalytic().

_visibilities: Vectorized version of visibility_point if you can gain additional speed

intensitymap: Computes the whole image of the model

intensitymap!: Inplace version of intensitymap

ComradeBase.AbstractPolarizedModel — Type

abstract type AbstractPolarizedModel <: ComradeBase.AbstractModel

Type the classifies a model as being intrinsically polarized. This means that any call to visibility must return a StokesParams to denote the full stokes polarization of the model.

ComradeBase.CirBasis — Type

CirBasis <: PolBasis

Measurement uses the circular polarization basis, which is typically used for circular feed interferometers.

ComradeBase.CoherencyMatrix — Type

CoherencyMatrix(s::StokesParams, basis1::PolBasis)
CoherencyMatrix(s::StokesParams, basis1::PolBasis, basis2::PolBasis)
CoherencyMatrix(s::StokesParams, basis1::PolBasis, basis2::PolBasis, refbasis=CirBasis())

Constructs the coherency matrix from the set of stokes parameters s. This is specialized on basis1 and basis2 which form the tensor product basis |basis1><basis2|, or if a single basis is given then by |basis><basis|.

For example

CoherencyMatrix(s, CircBasis())

will give the coherency matrix

|I+V Q+iU| |Q-iU I-V |

while

CoherencyMatrix(s, LinBasis())

will give |I+Q U+iV| |U-iV I-Q |

Notes

Internally this function first converts to a reference basis and then the final basis. You can select the reference basis used with the optional argument refbasis. By default we use the circular basis as our reference. Note that this is only important for mixed bases, e.g., if basis1 and basis2 are different. If basis1==basis2 then the reference basis is never used.

ComradeBase.CoherencyMatrix — Type

struct CoherencyMatrix{B1, B2, T} <: StaticArraysCore.FieldMatrix{2, 2, T}

Coherency matrix for a single baseline with bases B1 and B2. The two bases correspond to the type of feeds used for each telescope and should be subtypes of PolBasis. To see which bases are implemented type subtypes(Rimes.PolBasis) in the REPL.

For a circular basis the layout of the coherency matrix is

RR* RL*
LR* RR*

which can be constructed using

c = CoherencyMatrix(RR, LR, RL, LL, CirBasis())

For a linear basis the layout of the coherency matrix is

XX* XY*
YX* YY*

which can be constructed using

c = CoherencyMatrix(XX, YX, XY, YY, CirBasis())

For a mixed (e.g., circular and linear basis) the layout of the coherency matrix is

RX* RY*
LX* LY*

or e.g., linear and circular the layout of the coherency matrix is

XR* XL*
YR* YL*

These coherency matrices can be constructed using:

# Circular and linear feeds i.e., |R><X|
c = CoherencyMatrix(RX, LX, RY, LY, LinBasis(), CirBasis())
# Linear and circular feeds i.e., |X><R|
c = CoherencyMatrix(XR, YR, XL, YL, LinBasis(), CirBasis())

ComradeBase.CoherencyMatrix — Method

CoherencyMatrix(e11, e21, e12, e22, basis1::PolBasis basis2::PolBasis)

Constructs the coherency matrix with components e11 e12 e21 e22 relative to the tensor product basis, basis given by |basis1><basis2|.

For instance

c = Coherency(1.0, 0.0, 0.0, 1.0, CirBasis(), LinBasis())

elements correspond to RX* RY* LX* LY*

ComradeBase.CoherencyMatrix — Method

CoherencyMatrix(e11, e21, e12, e22, basis::PolBasis)

Constructs the coherency matrix with components e11 e12 e21 e22 relative to the tensor product basis, basis given by |basis><basis|.

For instance

c = Coherency(1.0, 0.0, 0.0, 1.0, CirBasis())

elements correspond to RR* RL* LR* LL*

ComradeBase.CoherencyMatrix — Method

CoherencyMatrix(e11, e21, e12, e22, basis::NTuple{2, PolBasis})

Constructs the coherency matrix with components e11 e12 e21 e22 relative to the tensor product basis, |basis[1]><basis[2]|. Note that basis[1] and basis[2] could be different.

For instance

c = Coherency(1.0, 0.0, 0.0, 1.0, CirBasis(), LinBasis())

elements correspond to RX* RY* LX* LY*

ComradeBase.DensityAnalytic — Type

DensityAnalytic

Internal type for specifying the nature of the model functions. Whether they can be easily evaluated pointwise analytic. This is an internal type that may change.

ComradeBase.ElectricFieldBasis — Type

abstract type ElectricFieldBasis

An abstract type whose subtypes denote a specific electric field basis.

ComradeBase.GriddedKeys — Type

struct GriddedKeys{N, G, Hd, T} <: ComradeBase.AbstractDims{N, T}

This struct holds the dimensions that the EHT expect. The first type parameter N defines the names of each dimension. These names are usually one of - (:X, :Y, :T, :F) - (:X, :Y, :F, :T) - (:X, :Y) # spatial only where :X,:Y are the RA and DEC spatial dimensions respectively, :T is the the time direction and :F is the frequency direction.

Fieldnames

dims

header

Notes

Warning it is rare you need to access this constructor directly. Instead use the direct IntensityMap function.

ComradeBase.GriddedKeys — Method

GriddedKeys(dims::NamedTuple{Na}, header=nothing)

Builds the EHT image dimensions using the names Na and dimensions are the values of dims. You can also optionally has a header that stores additional information from e.g., a FITS header. The type parameter Na defines the names of each dimension. These names are usually one of - (:X, :Y, :T, :F) - (:X, :Y, :F, :T) - (:X, :Y) # spatial only where :X,:Y are the RA and DEC spatial dimensions respectively, :T is the the time direction and :F is the frequency direction.

Notes

Instead use the direct IntensityMap function.

dims = GriddedKeys((X=-5.0:0.1:5.0, Y=-4.0:0.1:4.0, T=[1.0, 1.5, 1.75], F=[230, 345]))

ComradeBase.GriddedKeys — Method

GriddedKeys{Na}(dims::Tuple, header=nothing) where {Na}

Builds the EHT image dimensions using the names Na and dimensions dims. You can also optionally has a header that stores additional information from e.g., a FITS header. The type parameter Na defines the names of each dimension. These names are usually one of - (:X, :Y, :T, :F) - (:X, :Y, :F, :T) - (:X, :Y) # spatial only where :X,:Y are the RA and DEC spatial dimensions respectively, :T is the the time direction and :F is the frequency direction.

Notes

Instead use the direct IntensityMap function.

dims = GriddedKeys{(:X, :Y, :T, :F)}((-5.0:0.1:5.0, -4.0:0.1:4.0, [1.0, 1.5, 1.75], [230, 345]))

ComradeBase.IntensityMap — Method

IntensityMap(data::AbstractArray, dims::NamedTuple)
IntensityMap(data::AbstractArray, grid::AbstractDims)

Constructs an intensitymap using the image dimensions given by dims. This returns a KeyedArray with keys given by an ImageDimensions object.

dims = (X=range(-10.0, 10.0, length=100), Y = range(-10.0, 10.0, length=100),
        T = [0.1, 0.2, 0.5, 0.9, 1.0], F = [230e9, 345e9]
        )
imgk = IntensityMap(rand(100,100,5,1), dims)

ComradeBase.IsAnalytic — Type

struct IsAnalytic <: ComradeBase.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has a analytic fourier transform and/or image.

ComradeBase.IsPrimitive — Type

struct IsPrimitive

Trait for primitive model

ComradeBase.LPol — Type

struct LPol <: ComradeBase.ElectricFieldBasis

The left circular electric field basis, i.e. a left-handed circular feed.

ComradeBase.LinBasis — Type

LinBasis <: PolBasis

Measurement uses the linear polarization basis, which is typically used for linear feed interferometers.

ComradeBase.NotAnalytic — Type

struct NotAnalytic <: ComradeBase.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has does not have a easy analytic fourier transform and/or intensity function.

ComradeBase.NotPrimitive — Type

struct NotPrimitive

Trait for not-primitive model

ComradeBase.PolBasis — Type

struct PolBasis{B1<:Union{Missing, ComradeBase.ElectricFieldBasis}, B2<:Union{Missing, ComradeBase.ElectricFieldBasis}}

Denotes a general polarization basis, with basis vectors (B1,B2) which are typically <: Union{ElectricFieldBasis, Missing}

ComradeBase.PrimitiveTrait — Type

abstract type PrimitiveTrait

This trait specifies whether the model is a primitive

Notes

This will likely turn into a trait in the future so people can inject their models into Comrade more easily.

ComradeBase.RPol — Type

struct RPol <: ComradeBase.ElectricFieldBasis

The right circular electric field basis, i.e. a right-handed circular feed.

ComradeBase.SingleStokes — Type

struct SingleStokes{M, S} <: ComradeBase.AbstractModel

Helper function that converts a model from something that compute polarized images to just a single stokes parameter. This is useful if you just want to fit a single stokes parameter.

ComradeBase.StokesIntensityMap — Type

struct StokesIntensityMap{T, N, SI, SQ, SU, SV}

General struct that holds intensity maps for each stokes parameter. Each image I, Q, U, V must share the same axis dimensions. This type also obeys much of the usual array interface in Julia. The following methods have been implemented:

size

eltype (returns StokesParams)

ndims

getindex

setindex!

pixelsizes

fieldofview

imagepixels

imagegrid

stokes

Warning

This may eventually be phased out for IntensityMaps whose base types are StokesParams, but currently we use this for speed reasons with Zygote.

ComradeBase.StokesParams — Type

struct StokesParams{T} <: StaticArraysCore.FieldVector{4, T}

Static vector that holds the stokes parameters of a polarized complex visibility

To convert between a StokesParams and CoherencyMatrix use the convert function

convert(::CoherencyMatrix, StokesVector(1.0, 0.1, 0.1, 0.4))

ComradeBase.XPol — Type

struct XPol <: ComradeBase.ElectricFieldBasis

The horizontal or X electric feed basis, i.e. the horizontal linear feed.

ComradeBase.YPol — Type

struct YPol <: ComradeBase.ElectricFieldBasis

The vertical or Y electric feed basis, i.e. the vertical linear feed.

ComradeBase.axisdims — Method

axisdims(img::IntensityMap)

Returns the keys of the IntensityMap as the actual internal AbstractDims object.

ComradeBase.basis_components — Function

basis_components([T=Float64,], e::ElectricFieldBasis, b::PolBasis)

Returns a static vector that contains the components of the electric field basis vector e in terms of the polarization basis b. The first argument is optionally the eltype of the static vector.

Examples

julia> basis_components(Float64, R(), PolBasis{XPol,Y}())
2-element StaticArraysCore.SVector{2, ComplexF64} with indices SOneTo(2):
 0.7071067811865475 + 0.0im
                0.0 - 0.7071067811865475im

julia> basis_components(R(), PolBasis{XPol,Y}())
2-element StaticArraysCore.SVector{2, ComplexF64} with indices SOneTo(2):
 0.7071067811865475 + 0.0im
                0.0 - 0.7071067811865475im

julia> basis_components(Float64, X(), PolBasis{XPol,Y}())
2-element StaticArraysCore.SVector{2, ComplexF64} with indices SOneTo(2):
 1.0 + 0.0im
 0.0 + 0.0im

ComradeBase.basis_transform — Function

basis_transform([T=Float64,], b1::PolBasis, b2::PolBasis)
basis_transform([T=Float64,], b1::PolBasis=>b2::PolBasis)

Produces the transformation matrix that transforms the vector components from basis b1 to basis b2. This means that if for example E is the circular basis then basis_transform(CirBasis=>LinBasis)E is in the linear basis. In other words the columns of the transformation matrix are the coordinate vectors of the new basis vectors in the old basis.

Example

julia> basis_transform(CirBasis()=>LinBasis())
2×2 StaticArraysCore.SMatrix{2, 2, ComplexF64, 4} with indices SOneTo(2)×SOneTo(2):
 0.707107-0.0im       0.707107-0.0im
      0.0-0.707107im       0.0+0.707107im

ComradeBase.centroid — Method

centroid(im::AbstractIntensityMap)

Computes the image centroid aka the center of light of the image.

ComradeBase.dims — Method

dims(g::AbstractDims)

Returns a tuple containing the dimensions of g. For a named version see named_dims

ComradeBase.evpa — Method

evpa(m)

Compute the evpa of a stokes vector or cohereny matrix.

ComradeBase.fieldofview — Method

fieldofview(img::IntensityMap)
fieldofview(img::IntensityMapTypes)

Returns a named tuple with the field of view of the image.

ComradeBase.flux — Method

flux(im::IntensityMap)
flux(img::StokesIntensityMap)

Computes the flux of a intensity map

ComradeBase.header — Method

header(img::IntensityMap)

Retrieves the header of an IntensityMap

ComradeBase.header — Method

header(g::AbstractDims)

Returns the headerinformation of the dimensions g

ComradeBase.imagegrid — Method

imagegrid(k::IntensityMap)

Returns the grid the IntensityMap is defined as. Note that this is unallocating since it lazily computes the grid. The grid is an example of a KeyedArray and works similarly. This is useful for broadcasting a model across an abritrary grid.

ComradeBase.imagepixels — Method

imagepixels(img::IntensityMap)
imagepixels(img::IntensityMapTypes)

Returns a abstract spatial dimension with the image pixels locations X and Y.

ComradeBase.imanalytic — Method

imanalytic(::Type{<:AbstractModel})

Determines whether the model is pointwise analytic in the image domain, i.e. we can evaluate its intensity at an arbritrary point.

If IsAnalytic() then it will try to call intensity_point to calculate the intensity.

ComradeBase.innerprod — Function

innerprod(::Type{T}, XPol(), YPol())

Computes the complex inner product of two elements of a complex Hilbert space X and Y where base element of the output is T.

ComradeBase.intensity_point — Function

intensity_point(model::AbstractModel, p)

Function that computes the pointwise intensity if the model has the trait in the image domain IsAnalytic(). Otherwise it will use construct the image in visibility space and invert it.

ComradeBase.intensitymap — Function

intensitymap(s, fovx, fovy, nx, ny, x0=0.0, y0=0.0)

Creates a spatial only IntensityMap intensity map whose pixels in the x, y direction are such that the image has a field of view fovx, fovy, with the number of pixels nx, ny, and the origin or phase center of the image is at x0, y0.

ComradeBase.intensitymap — Function

intensitymap(model::AbstractModel, args...)

Computes the intensity map of model. For the inplace version see intensitymap!

ComradeBase.intensitymap! — Function

intensitymap!(buffer::AbstractDimArray, model::AbstractModel)

Computes the intensity map of model by modifying the buffer

ComradeBase.intensitymap! — Method

intensitymap!(img::AbstractIntensityMap, mode;, executor = SequentialEx())

Computes the intensity map or image of the model. This updates the IntensityMap object img.

Optionally the user can specify the executor that uses FLoops.jl to specify how the loop is done. By default we use the SequentialEx which uses a single-core to construct the image.

ComradeBase.intensitymap — Method

intensitymap(model::AbstractModel, dims)

Computes the intensity map or image of the model. This returns an IntensityMap which is a KeyedArray with ImageDimensions as keys. The dimensions are a NamedTuple and must have one of the following names: - (:X, :Y, :T, :F) - (:X, :Y, :F, :T) - (:X, :Y) # spatial only where :X,:Y are the RA and DEC spatial dimensions respectively, :T is the the time direction and :F is the frequency direction.

ComradeBase.ispolarized — Method

ispolarized(::Type)

Trait function that defines whether a model is polarized or not.

ComradeBase.isprimitive — Function

isprimitive(::Type)

Dispatch function that specifies whether a type is a primitive Comrade model. This function is used for dispatch purposes when composing models.

Notes

If a user is specifying their own model primitive model outside of Comrade they need to specify if it is primitive

struct MyPrimitiveModel end
ComradeBase.isprimitive(::Type{MyModel}) = ComradeBase.IsPrimitive()

ComradeBase.linearpol — Method

linearpol(s)

Computes linearpol from a set of stokes parameters s.

ComradeBase.m̆ — Method

m̆(m)

Compute the fractional linear polarization of a stokes vector or coherency matrix

ComradeBase.named_dims — Method

named_dims(g::AbstractDims)

Returns a named tuple containing the dimensions of g. For a unnamed version see dims

ComradeBase.phasecenter — Method

phasecenter(img::IntensityMap)
phasecenter(img::StokesIntensitymap)

Computes the phase center of an intensity map. Note this is the pixels that is in the middle of the image.

ComradeBase.pixelsizes — Method

pixelsizes(img::IntensityMap)
pixelsizes(img::IntensityMapTypes)

Returns a named tuple with the spatial pixel sizes of the image.

ComradeBase.radialextent — Function

radialextent(model::AbstractModel)

Provides an estimate of the radial size/extent of the model. This is used internally to estimate image size when plotting and using modelimage

ComradeBase.second_moment — Method

second_moment(im::AbstractIntensityMap; center=true)

Computes the image second moment tensor of the image. By default we really return the second cumulant or centered second moment, which is specified by the center argument.

ComradeBase.second_moment — Method

centroid(im::AbstractIntensityMap)

Computes the image centroid aka the center of light of the image.

ComradeBase.stokes — Method

stokes(m::AbstractPolarizedModel, p::Symbol)

Extract the specific stokes component p from the polarized model m

ComradeBase.visanalytic — Method

visanalytic(::Type{<:AbstractModel})

Determines whether the model is pointwise analytic in Fourier domain, i.e. we can evaluate its fourier transform at an arbritrary point.

If IsAnalytic() then it will try to call visibility_point to calculate the complex visibilities. Otherwise it fallback to using the FFT that works for all models that can compute an image.

ComradeBase.visibility_point — Function

visibility_point(model::AbstractModel, p)

Function that computes the pointwise visibility. This must be implemented in the model interface if visanalytic(::Type{MyModel}) == IsAnalytic()