ComradeBase

Documentation for ComradeBase.

ComradeBase.BSplinePulse

ComradeBase.DeltaPulse

ComradeBase.DensityAnalytic

ComradeBase.IsAnalytic

ComradeBase.NotAnalytic

ComradeBase.PrimitiveTrait

ComradeBase.Pulse

ComradeBase.SingleStokes

ComradeBase.SqExpPulse

ComradeBase.flux

ComradeBase.imanalytic

ComradeBase.intensity_point

ComradeBase.intensitymap

ComradeBase.intensitymap!

ComradeBase.isprimitive

ComradeBase.visanalytic

ComradeBase.visibility_point

ComradeBase.BSplinePulse — Type

$(TYPEDEF)

Uses the basis spline (BSpline) kernel of order N. These are the kernel that come from recursively convolving the tophat kernel

\[ B_0(x) = \begin{cases} 1 & |x| < 1 \\ 0 & otherwise \end{cases}\]

N times.

Notes

BSpline kernels have a number of nice properties:

Simple frequency response $\sinc(u/2)^N$

preserve total intensity

For N>1 these kernels aren't actually interpolation kernels however, this doesn't matter for us.

Currently only the 0,1,3 order kernels are implemented.

ComradeBase.DeltaPulse — Type

struct DeltaPulse{T} <: ComradeBase.Pulse

A dirac comb pulse function. This means the image is just the dicrete Fourier transform

ComradeBase.DensityAnalytic — Type

DensityAnalytic

Internal type for specifying the nature of the model functions. Whether they can be easily evaluated pointwise analytic. This is an internal type that may change.

ComradeBase.IsAnalytic — Type

struct IsAnalytic <: ComradeBase.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has a analytic fourier transform and/or image.

ComradeBase.NotAnalytic — Type

struct NotAnalytic <: ComradeBase.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has does not have a easy analytic fourier transform and/or intensity function.

ComradeBase.PrimitiveTrait — Type

abstract type PrimitiveTrait

This trait specifies whether the model is a primitive

Notes

This will likely turn into a trait in the future so people can inject their models into Comrade more easily.

ComradeBase.Pulse — Type

Pulse Pixel response function for a radio image model. This makes a discrete sampling continuous by picking a certain smoothing kernel for the image.

Notes

To see the implemented Pulses please use the subtypes function i.e. subtypes(Pulse)

ComradeBase.SingleStokes — Type

struct SingleStokes{M, S} <: ComradeBase.AbstractModel

Helper function that converts a model from something that compute polarized images to just a single stokes parameter. This is useful if you just want to fit a single stokes parameter.

ComradeBase.SqExpPulse — Type

struct SqExpPulse{T} <: ComradeBase.Pulse

Normalized square exponential kernel, i.e. a Gaussian. Note the smoothness is modfied with ϵ which is the inverse variance in units of 1/pixels².

ComradeBase.flux — Method

flux(im)

Computes the flux of a intensity map

ComradeBase.imanalytic — Method

imanalytic(::Type{<:AbstractModel})

Determines whether the model is pointwise analytic in the image domain, i.e. we can evaluate its intensity at an arbritrary point.

If IsAnalytic() then it will try to call intensity_point to calculate the intensity.

ComradeBase.intensity_point — Function

Function that computes the pointwise intensity if the model has the trait in the image domain IsAnalytic(). Otherwise it will use construct the image in visibility space and invert it.

ComradeBase.intensitymap — Function

Computes the intensity map of model. This version requires additional information to construct the grid.

Example

m = Gaussian()
# field of view
fovx, fovy = 5.0
fovy = 5.0
# number of pixels
nx, ny = 128

img = intensitymap(m, fovx, fovy, nx, ny; pulse=DeltaPulse())

ComradeBase.intensitymap! — Function

Computes the intensity map of model by modifying the input IntensityMap object

ComradeBase.isprimitive — Function

isprimitive(::Type)

Dispatch function that specifies whether a type is a primitive Comrade model. This function is used for dispatch purposes when composing models.

Notes

If a user is specifying their own model primitive model outside of Comrade they need to specify if it is primitive

struct MyPrimitiveModel end
Comrade.isprimitive(::Type{MyModel}) = Comrade.IsPrimitive()

ComradeBase.visanalytic — Method

visanalytic(::Type{<:AbstractModel})

Determines whether the model is pointwise analytic in Fourier domain, i.e. we can evaluate its fourier transform at an arbritrary point.

If IsAnalytic() then it will try to call visibility_point to calculate the complex visibilities. Otherwise it fallback to using the FFT that works for all models that can compute an image.

ComradeBase.visibility_point — Function

Function that computes the pointwise visibility if the model has the trait in the fourier domain IsAnalytic(). Otherwise it will use the FFTW fallback.